pyproject.toml
[tool.poetry]
name = "cows-tsp"
version = "0.1.0"
description = ""
authors = ["Jason Chen <chezbgone@gmail.com>"]
readme = "README.md"
packages = [{include = "cows_tsp"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.0"
more-itertools = "^9.1.0"
python-tsp = "^0.3.1"

[tool.pyright]
venvPath = "/home/jason/.cache/pypoetry/virtualenvs"
venv = "cows-tsp-I_ZEdn1U-py3.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
